# spec/openapi.yaml
openapi: 3.0.1
info:
  description: An API suite providing read only access to search and retrieve
    public company data
  title: Companies House Public Data API
  version: 1.0.0
servers:
  - url: https://api.company-information.service.gov.uk/
  - url: http://api.company-information.service.gov.uk/
security:
  - api_key: []
...
paths:
...
  /search/disqualified-officers:
    get:
      description: Search for disqualified officer information
      operationId: searchDisqualifiedOfficersGET
      parameters:
        - description: The term being searched for.
          in: query
          name: q
          required: true
          schema:
            type: string
        - description: The number of search results to return per page.
          in: query
          name: items_per_page
          schema:
            type: integer
        - description: The index of the first result item to return.
          in: query
          name: start_index
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        description_identifiers:
                          type: array
                          items:
                            type: string
                        description:
                          type: string
                        snippet:
                          type: string
                        address_snippet:
                          type: string
                        date_of_birth:
                          type: string
                        title:
                          type: string
                        links:
                          type: object
                          properties:
                            self:
                              type: string
                          required:
                            - self
                        address:
                          type: object
                          properties:
                            country:
                              type: string
                            premises:
                              type: string
                            locality:
                              type: string
                            postal_code:
                              type: string
                            address_line_1:
                              type: string
                            address_line_2:
                              type: string
                            region:
                              type: string
                          required:
                            - locality
                        matches:
                          type: object
                          properties:
                            snippet:
                              type: array
                            title:
                              type: array
                              items:
                                type: integer
                          required:
                            - snippet
                            - title
                        kind:
                          type: string
                      required:
                        - description
                        - address_snippet
                        - title
                        - links
                        - address
                        - kind
                  items_per_page:
                    type: integer
                  kind:
                    type: string
                  page_number:
                    type: integer
                  start_index:
                    type: integer
                  total_results:
                    type: integer
                required:
                  - items
                  - items_per_page
                  - kind
                  - page_number
                  - start_index
                  - total_results
                additionalProperties: false
                title: searchDisqualifiedOfficers
                example:
                  items: []
                  items_per_page: 20
                  kind: search#disqualified-officers
                  page_number: 1
                  start_index: 0
                  total_results: 0
          description: Search all
        "401":
          content: {}
          description: Not authorised
      summary: Search disqualified officers
      tags:
        - search
      x-operationName: search disqualified officers
      x-eov-operation-handler: controllers/SearchController
...
components:
...
  securitySchemes:
    api_key:
      in: header
      name: api_key
      type: apiKey
...